addlogin

+++ application/admin/controller/Index.php
@@ -16,7 +16,16 @@ class Index extends Base
 
         if ($response['status']){
             $res = $response['data']->addGoodsItem($post);
-            if (request()->file('image')){
+            if (request()->file('image') && $res){
+                $arr = Upload::upload_file(request()->file('image'))->getData();
+                if ($arr['status']==200 && $res){
+                    $data['index_id'] = $res['id'];
+                    $data['url'] = $arr['data'];
+                    $user = new Image();
+                    $user->save($data);
+                }
+            }
+            if (request()->file('images') && $res){
                 $arr = Upload::upload_file()->getData();
                 if ($arr['status']==200 && $res){
                     $data['index_id'] = $res['id'];
@@ -25,7 +34,7 @@ class Index extends Base
                     $user->save($data);
                 }
             }
-            if (request()->file('move')){
+            if (request()->file('move') && $res){
                 $arr = Upload::upload_file()->getData();
                 if ($arr['status']==200 && $res){
                     $data['index_id'] = $res['id'];
@@ -43,14 +52,14 @@ class Index extends Base
     public function update ()
     {
         $response = self::common([
-            'Validate'   =>  'Index.update',
+            'Validate'   =>  'Index.change',
             'Model'      =>  'Index'
         ]);
         $post = Request::param();
 
         if ($response['status']){
             $res = $response['data']->changeGoodsItem($post);
-            if (request()->file('image')){
+            if (request()->file('image') && $res){
                 $user = new Image();
                 $img = $user->where('index_id='.$post['id'])->find();
                 unlink('.'.$img['url']);
@@ -60,7 +69,7 @@ class Index extends Base
                     $user->where('index_id='.$post['id'])->update($data);
                 }
             }
-            if (request()->file('move')){
+            if (request()->file('move') && $res){
                 $user = new Image();
                 $img = $user->where('index_id='.$post['id'])->find();
                 unlink('.'.$img['url']);
@@ -75,4 +84,32 @@ class Index extends Base
         return $response['data'];
 
     }
+    public function lists ()
+    {
+        $response = self::common([
+            'Validate'   =>  'Index.type',
+            'Model'      =>  'Index'
+        ]);
+        return false === $response['status'] ? $response['data'] : $response['data']->lists(Request::param());
+    }
+    public function del ()
+    {
+        $respone = self::common([
+            'Validate'   =>  'Index.change',
+            'Model'      =>  'Index'
+        ]);
+
+        return false === $respone['status'] ? $respone['data'] : $respone['data']->deleteGoodsItem(Request::param());
+    }
+
+    public function up ()
+    {
+        $respone = self::common([
+            'Validate'   =>  'Index.change',
+            'Model'      =>  'Index'
+        ]);
+
+        return false === $respone['status'] ? $respone['data'] : $respone['data']->upGoodsItem(Request::param());
+    }
+
 }
